requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
chromedriver-binary==84.0.4147.30.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.0
progressbar2==3.51.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-utils==2.4.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
